Add LDAP group travelling_accounts_jira (#780)
We are in a limbo state where Security has begun allowing travelling accounts to access Jira, but at the same time we don't have a solid policy on allowing EVERY travelling account to access Jira.  This is a stand-in group so that we can easily allow this.  If a policy change happens, we can s/_jira// later.